pyproject.toml
[tool.poetry]
name = "scrap-a-job"
version = "0.1.0"
description = "Find a job or learn about employment market with data-scraping"
authors = ["NanoClem <clemdecoop.pro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.1"
requests = "^2.28.1"
pyaml = "^21.10.1"
Scrapy = "^2.6.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
